reconfigure ports
