+++ apps/docs/package.json
@@ -1,31 +1,31 @@
 {
-	"name": "docs",
-	"version": "1.0.0",
-	"private": true,
-	"scripts": {
-		"dev": "next dev --port 3001",
-		"build": "next build",
-		"start": "next start",
-		"lint": "next lint"
-	},
-	"dependencies": {
-		"@repo/ui": "workspace:*",
-		"next": "latest",
-		"react": "latest",
-		"react-dom": "latest"
-	},
-	"devDependencies": {
-		"@next/eslint-plugin-next": "latest",
-		"@repo/eslint-config": "workspace:*",
-		"@repo/typescript-config": "workspace:*",
-		"@repo/tailwind-config": "workspace:*",
-		"@types/eslint": "latest",
-		"@types/node": "latest",
-		"@types/react": "latest",
-		"@types/react-dom": "latest",
-		"autoprefixer": "latest",
-		"eslint": "latest",
-		"tailwindcss": "latest",
-		"typescript": "latest"
-	}
+  "name": "docs",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev --port 3002",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
+  },
+  "dependencies": {
+    "@repo/ui": "workspace:*",
+    "next": "latest",
+    "react": "latest",
+    "react-dom": "latest"
+  },
+  "devDependencies": {
+    "@next/eslint-plugin-next": "latest",
+    "@repo/eslint-config": "workspace:*",
+    "@repo/typescript-config": "workspace:*",
+    "@repo/tailwind-config": "workspace:*",
+    "@types/eslint": "latest",
+    "@types/node": "latest",
+    "@types/react": "latest",
+    "@types/react-dom": "latest",
+    "autoprefixer": "latest",
+    "eslint": "latest",
+    "tailwindcss": "latest",
+    "typescript": "latest"
+  }
 }
