+++ apps/operator/package.json
@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev -p 3001",
     "build": "next build",
-    "start": "next start",
+    "start": "next start -p 3001",
     "lint": "next lint"
   },
   "dependencies": {
@@ -33,6 +33,7 @@
     "eslint": "latest",
     "postcss": "latest",
     "tailwindcss": "latest",
+    "tailwind-variants": "latest",
     "typescript": "latest"
   }
 }
