+++ apps/operator/src/fonts/FTRegolaNeueTrial-Medium.woff2
