+++ apps/operator/src/fonts/FTRegolaNeueTrial-Regular.woff2
