+++ apps/operator/src/fonts/FTRegolaNeueTrial-Semibold.woff2
