+++ apps/operator/tsconfig.json
@@ -5,7 +5,10 @@
       {
         "name": "next"
       }
-    ]
+    ],
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   },
   "include": [
     "next-env.d.ts",
@@ -14,9 +17,7 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     "postcss.config.js",
-    "../../packages/config-typescript/typings.d.ts",
+    "../../packages/config-typescript/typings.d.ts"
   ],
-  "exclude": [
-    "node_modules"
-  ]
+  "exclude": ["node_modules"]
 }
